datasource-hub-rag-mcp/requirements.txt
fastmcp>=0.1.0
pymilvus>=2.4.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
